requirements.txt
python-dotenv
pyyaml
tqdm
beautifulsoup4
huggingface_hub
robotframework
robotframework-robocop